does not work well and the code is shit
This stopped working and i do not know what exactly i did that
made it stop working. But i my final exam next week and i can't
continue now, i have to learn for exams. I'll get back to this
clean up the code and then debug thoroughly to see what the
fuck is going on