clean up code and better error messages
did quite alot, can't quite remember some of the things that go
changed in this new commit. Yep i changed the toolchain from
clang with musl_libc to riscv64-gcc with newlib which is more
for embedded systems and that kind of stuff. Less syscalls to
implement and more important things to concentrate on.