makefile and cmdline opts for easy build/test/debug cycle
this commit features the ability to pass commandline options to the
emulator to tweak the runtime environment that it provides.
it also features a major improvement in the documentation of
the project. most important improvement is the ability to build
and run the emulator using make. No other build scripts are
necessary and you don't have to manually run `go generate` or
`go get`. In the future I'll extend the makefile to build for
other operating systems like bsd, mac and windows.

The improvement is supposed to make it easier for people to get
started if they want to start hacking on the project.